+++ src/mocks/handlers/index.ts
@@ -1,0 +1,3 @@
+export { cartsHandlers } from './carts';
+export { ordersHandler } from './orders';
+export { productsHandler } from './products';
