+++ src/mocks/index.ts
@@ -1,0 +1,7 @@
+import { cartsHandlers, ordersHandler, productsHandler } from './handlers';
+
+export const handlers = [
+  ...cartsHandlers,
+  ...ordersHandler,
+  ...productsHandler
+];
