+++ src/mocks/worker.ts
@@ -1,4 +1,4 @@
 import { setupWorker } from 'msw';
-import { handlers } from './handlers';
+import { handlers } from './index';
 
 export const worker = setupWorker(...handlers);
